performance improvements
