+++ packages/suite-base/src/components/Sidebars/NewSidebar.tsx
@@ -6,85 +6,15 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import CloseIcon from "@mui/icons-material/Close";
-import { Badge, BadgeProps, Divider, IconButton, Tab, Tabs } from "@mui/material";
-import { makeStyles } from "tss-react/mui";
+import { Badge, Divider, IconButton, Tab, Tabs } from "@mui/material";
+import { useCallback, useMemo } from "react";
 
+import { useStyles } from "@lichtblick/suite-base/components/Sidebars/NewSidebar.style";
+import { NewSidebarItem } from "@lichtblick/suite-base/components/Sidebars/types";
 import Stack from "@lichtblick/suite-base/components/Stack";
 
-const useStyles = makeStyles()((theme) => ({
-  root: {
-    boxSizing: "content-box",
-    backgroundColor: theme.palette.background.paper,
-  },
-  badgeRoot: {
-    display: "flex",
-    alignItems: "baseline",
-    gap: theme.spacing(1),
-  },
-  badge: {
-    fontSize: theme.typography.caption.fontSize,
-    padding: theme.spacing(0.125, 0.75),
-    borderRadius: 8,
-    transform: "none",
-    position: "relative",
-  },
-  badgeInvisible: {
-    display: "none",
-  },
-  anchorRight: {
-    borderLeft: `1px solid ${theme.palette.divider}`,
-  },
-  anchorLeft: {
-    borderRight: `1px solid ${theme.palette.divider}`,
-  },
-  tabs: {
-    minHeight: "auto",
-    flex: "1 1 auto",
-    overflow: "hidden",
-    paddingLeft: theme.spacing(0.25),
-
-    ".MuiTabs-indicator": {
-      transform: "scaleX(0.5)",
-      height: 2,
-    },
-    ".MuiTab-root": {
-      minHeight: 30,
-      minWidth: theme.spacing(4),
-      padding: theme.spacing(0, 1),
-      color: theme.palette.text.secondary,
-      fontSize: "0.6875rem",
-
-      "&.Mui-selected": {
-        color: theme.palette.text.primary,
-      },
-    },
-  },
-  iconButton: {
-    padding: theme.spacing(0.91125), // round out the overall height to 30px
-    color: theme.palette.text.secondary,
-    borderRadius: 0,
-
-    ":hover": {
-      color: theme.palette.text.primary,
-    },
-  },
-  tabContent: {
-    flex: "auto",
-    overflow: "auto",
-  },
-}));
-
-export type NewSidebarItem = {
-  title: string;
-  component: React.ComponentType;
-  badge?: {
-    color: BadgeProps["color"];
-    count: number;
-  };
-};
-
-function Noop(): ReactNull {
-  return ReactNull;
+function Noop(): JSX.Element | undefined {
+  return undefined;
 }
 
 export function NewSidebar<K extends string>({
@@ -102,7 +32,42 @@ export function NewSidebar<K extends string>({
 }): JSX.Element {
   const { classes, cx } = useStyles();
 
-  const SelectedComponent = (activeTab != undefined && items.get(activeTab)?.component) ?? Noop;
+  const handleTabChange = useCallback(
+    (_ev: React.SyntheticEvent, newValue: K) => {
+      if (newValue !== activeTab) {
+        setActiveTab(newValue);
+      }
+    },
+    [activeTab, setActiveTab],
+  );
+
+  const SelectedComponent = useMemo(() => {
+    return (activeTab != undefined && items.get(activeTab)?.component) ?? Noop;
+  }, [activeTab, items]);
+
+  const memoizedTabs = useMemo(() => {
+    return Array.from(items.entries(), ([key, item]) => (
+      <Tab
+        key={key}
+        label={
+          <Badge
+            invisible={!item.badge}
+            badgeContent={item.badge?.count}
+            color={item.badge?.color}
+            classes={{
+              root: classes.badgeRoot,
+              badge: classes.badge,
+              invisible: classes.badgeInvisible,
+            }}
+          >
+            {item.title}
+          </Badge>
+        }
+        value={key}
+        data-testid={`${key}-${anchor}`}
+      />
+    ));
+  }, [items, classes.badgeRoot, classes.badge, classes.badgeInvisible, anchor]);
 
   return (
     <Stack
@@ -119,33 +84,9 @@ export function NewSidebar<K extends string>({
           className={classes.tabs}
           textColor="inherit"
           value={activeTab ?? false}
-          onChange={(_ev, newValue: K) => {
-            if (newValue !== activeTab) {
-              setActiveTab(newValue);
-            }
-          }}
+          onChange={handleTabChange}
         >
-          {Array.from(items.entries(), ([key, item]) => (
-            <Tab
-              key={key}
-              label={
-                <Badge
-                  invisible={item.badge == undefined}
-                  badgeContent={item.badge?.count}
-                  color={item.badge?.color}
-                  classes={{
-                    root: classes.badgeRoot,
-                    badge: classes.badge,
-                    invisible: classes.badgeInvisible,
-                  }}
-                >
-                  {item.title}
-                </Badge>
-              }
-              value={key}
-              data-testid={`${key}-${anchor}`}
-            />
-          ))}
+          {memoizedTabs}
         </Tabs>
 
         <IconButton
